[REFACTOR] #52 Hilt 모듈을 수명주기, 기능에 따라 분리
 - App -> Retrofit -> DataSource -> Service -> Repository 순으로 의존성이 주입되는 흐름
 - Service와 Repository는 ViewModel 생명주기 내에서 사용되니 ViewModelComponent 생명주기로 설
