+++ app/src/main/java/co/kr/woowahan_repo/di/RepositoryModule.kt
@@ -1,21 +1,16 @@
 package co.kr.woowahan_repo.di
 
-import co.kr.woowahan_repo.data.datasource.WoowahanSharedPreferences
 import co.kr.woowahan_repo.data.repository.*
-import co.kr.woowahan_repo.domain.datasource.GithubTokenDataSource
 import co.kr.woowahan_repo.domain.repository.*
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
+import dagger.hilt.android.components.ViewModelComponent
 
 
 @Module
-@InstallIn(SingletonComponent::class)
+@InstallIn(ViewModelComponent::class)
 abstract class RepositoryModule {
-
-    @Binds
-    abstract fun bindGithubTokenDataSource(impl: WoowahanSharedPreferences): GithubTokenDataSource
 
     @Binds
     abstract fun bindOAuthRepository(impl: GithubOAuthRepositoryImpl): GithubOAuthRepository
